Fix bug of giveall command
Currently, giveall allows users to add an assignment with same description but with a different due date from the existing in the data, resulting in different persons having same assignments but with different due dates.

Fix the giveall command to check if the input due date is same as the existing one.